Changed postController and added tests

+++ backend/controllers/postController.js
@@ -8,8 +8,18 @@ const Post = require('../model/Post.js')
 const postController = {
     async createPost(req, res) {
         try {
-            const {creatorId, forumId, postTitle, postText} = req.body;
-            
+            const {forumId, postTitle, postText} = req.body;
+
+            const creator_info = await User.findOne( {login: req.user.login} );
+
+            const creatorId = creator_info.uid;
+ 
+            const forum_info = await Forum.findOne( {forumId: forumId} );
+
+            if (!forum_info) {
+                return res.status(400).json( {error: "Forum with such ID does not exist"} );
+            }
+
             const counter_fetch = await Counter.findOne({_id: "Post"});
             const post_counter_value = counter_fetch.collectionCounter
             const post_number = post_counter_value + 1
@@ -48,21 +58,33 @@ const postController = {
         try {
             const {postId, forumId} = req.body;
 
-            const post = await Post.findOne({postId: postId});
-    
+            const post = await Post.findOne( {postId: postId}) ;
+
             if (!post) {
-                return res.status(400).json({error: "Post does not exists"})
+                return res.status(400).json( {error: "Post does not exists"} );
             }
     
-            const forum = await Forum.findOne({forumId: forumId});
+            const forum = await Forum.findOne( {forumId: forumId} );
     
             if (!forum) {
-                return res.status(400).json({error: "Forum does not exists"})
+                return res.status(400).json( {error: "Forum does not exists"} )
             }
 
-            forum.pinnedPost = postId;
-            await forum.save()
-            return res.status(200).json("Post pinned");
+            if (forum.forumId != post.forumId) {
+                return res.status(400).json( {error: "The post does not belong to the forum"} )
+            }
+
+            const user_info = await User.findOne( {login: req.user.login} );
+
+            const userId = user_info.uid;
+            const index = forum.moderatorIds.indexOf(userId);
+
+            if (index != -1 || userId == forum.creatorId) {
+                forum.pinnedPost = postId;
+                await forum.save()
+                return res.status(200).json( {message: "Post pinned"} );
+            }
+            return res.status(403).json( {error: "Unauthorized"} );
         } catch (error) {
             console.error("Error in pinPost:", error);
             res.status(500).json({ error: "Internal server error" });
@@ -91,13 +113,15 @@ const postController = {
         try {
             const {post_id} = req.params;
 
+            const delete_info = await User.findOne( {login: req.user.login} );
+
             const post_info = await Post.findOne( {postId: post_id} );
 
             if (!post_info) {
                 return res.status(400).json( {error: "post_id does not exist"} );
             }
 
-            const forum_id = post_info.forumId
+            const forum_id = post_info.forumId;
 
             const forum_info = await Forum.findOne( {forumId: forum_id} );
 
@@ -113,25 +137,32 @@ const postController = {
                 return res.status(400).json( {error: "The owner of the post does not exist"} );
             }
 
-            const user_post_ids = user_info.postIds;
-            const user_index = user_post_ids.indexOf(post_id);
+            const delete_id = delete_info.uid;
+            const index = forum_info.moderatorIds.indexOf(delete_id);
+
+            if (delete_id == forum_info.creatorId || index != -1 || delete_id == user_id) {
+
+                const user_post_ids = user_info.postsIds;
+                const user_index = user_post_ids.indexOf(post_id);
           
-            if (user_index != -1) {
-                user_post_ids.splice(user_index, 1);
+                if (user_index != -1) {
+                    user_post_ids.splice(user_index, 1);
+                }
+
+                const forum_post_ids = forum_info.postIds;
+                const forum_index = forum_post_ids.indexOf(post_id);
+
+                if (forum_index != -1) {
+                    forum_post_ids.splice(forum_index, 1);
+                }
+
+                await User.findOneAndUpdate( {uid: user_id}, {postIds: user_post_ids} ); // Remove the post entry from User
+                await Forum.findOneAndUpdate( {forumId: forum_id}, {postIds: forum_post_ids} ); // Remove the post entry from Forum
+                await Post.deleteOne( {postId: post_id} ); // Remove the post entry from database.
+
+                return res.status(200).json( {message: "Deleted a post successfuly"});
             }
-
-            const forum_post_ids = forum_info.postIds;
-            const forum_index = forum_post_ids.indexOf(post_id);
-
-            if (forum_index != -1) {
-                forum_post_ids.splice(forum_index, 1);
-            }
-
-            await User.findOneAndUpdate( {uid: user_id}, {postIds: user_post_ids} ); // Remove the post entry from User
-            await Forum.findOneAndUpdate( {forumId: forum_id}, {postIds: forum_post_ids} ); // Remove the post entry from Forum
-            await Post.deleteOne( {postId: post_id} ); // Remove the post entry from database.
-
-            return res.status(200).json( {message: "Deleted a post successfuly"});
+            return res.status(403).json( {error: "Unauthorized" });
 
         } catch (error) {
             console.error("Error in deletePost:", error);
